requirements.txt
discord.py==2.4.0
selenium==4.26.1
webdriver_manager==4.0.2
